+++ MonsterDB/Managers/SpawnSystem/SpawnManager.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using BepInEx;
 using BepInEx.Configuration;
 using HarmonyLib;
@@ -13,10 +14,14 @@ public static class SpawnManager
     private static readonly SpawnSystemList SpawnList;
     private static Dictionary<string, SpawnDataRef> spawnRefs;
     private static readonly CustomSyncedValue<string> sync;
+    private static readonly Dictionary<string, SpawnSystem.SpawnData> CachedSpawnData;
+    private static readonly Dictionary<string, BaseSpawnData> updateList;
     
     static SpawnManager()
     {
         spawnRefs = new Dictionary<string, SpawnDataRef>();
+        CachedSpawnData = new Dictionary<string, SpawnSystem.SpawnData>();
+        updateList = new Dictionary<string, BaseSpawnData>();
         sync = new CustomSyncedValue<string>(ConfigManager.ConfigSync, "MDB.ServerSync.SpawnList");
         SpawnList = MonsterDBPlugin.instance.gameObject.AddComponent<SpawnSystemList>();
         sync.ValueChanged += OnSyncChange;
@@ -25,6 +30,25 @@ public static class SpawnManager
     public static void Add(SpawnDataRef data)
     {
         spawnRefs[data.m_name] = data;
+    }
+
+    public static void QueueUpdate(BaseSpawnData data)
+    {
+        updateList[data.Prefab] = data;
+    }
+
+    public static List<string> GetCachedSpawnDataIds() => CachedSpawnData.Keys.ToList();
+    
+    public static bool TryGetSpawnData(string id, out SpawnSystem.SpawnData data) => CachedSpawnData.TryGetValue(id, out data);
+
+    public static void Export(string spawnerId)
+    {
+        if (!CachedSpawnData.TryGetValue(spawnerId, out SpawnSystem.SpawnData data)) return;
+        BaseSpawnData info = new BaseSpawnData();
+        info.SetupData(data);
+        string text = ConfigManager.Serialize(info);
+        string filePath = Path.Combine(FileManager.ExportFolder, info.Prefab + ".yml");
+        File.WriteAllText(filePath, text);
     }
 
     public static void Start()
@@ -76,7 +100,7 @@ public static class SpawnManager
             SpawnList.m_spawners.Add(spawnData);
         }
     }
-
+    
     public static void Update(SpawnDataRef info)
     {
         SpawnList.m_spawners.RemoveAll(x => x.m_name == info.m_name);
@@ -101,6 +125,26 @@ public static class SpawnManager
     {
         ValidateSpawns();
         __instance.m_spawnLists.Add(SpawnList);
+        CacheAndUpdateSpawners(__instance);
+    }
+
+    private static void CacheAndUpdateSpawners(SpawnSystem __instance)
+    {
+        List<SpawnSystemList>? lists = __instance.m_spawnLists;
+        for (int i = 0; i < lists.Count; ++i)
+        {
+            SpawnSystemList? list =  lists[i];
+            List<SpawnSystem.SpawnData>? spawners = list.m_spawners;
+            for (int j = 0; j < spawners.Count; ++j)
+            {
+                SpawnSystem.SpawnData? spawner = spawners[j];
+                CachedSpawnData[spawner.m_name] = spawner;
+                if (updateList.TryGetValue(spawner.m_name, out BaseSpawnData? data))
+                {
+                    data.UpdateData(spawner);
+                }
+            }
+        }
     }
 
     private static void ValidateSpawns()